direct_serial_subproject/requirements.txt
requests>=2.28.0
pyserial>=3.5
python-dotenv>=0.19.0
